requirements.txt
psutil
flask-cors
colorama
